backend/google-microservice/pyproject.toml
[tool.poetry]
name = "github-microservice"
version = "0.1.0"
description = "github-microservice"
authors = ["test"]
readme = "README.md"

package-mode = false


# Add production dependencies (ex: poetry add 'fastapi[standard]')
# Add dev dependencies (ex: poetry add ruff --group=dev)
[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pydantic-settings = "^2.5.2"
httpx = {extras = ["http2"], version = "^0.27.2"}
requests = "2.32.3"
google = "3.0.0"
google-cloud = "0.34.0"
google-cloud-pubsub = "2.26.0"
google-api-python-client = "2.14.0"
grpcio = ">=1.52.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
aiohttp = "^3.10.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

####################################
#       Custom ruff settings       #
####################################

[tool.ruff]
# Inherit exclude patterns from ruff.toml
extend-exclude = [".venv", ".env", "alembic"]

# Other settings from ruff.toml
fix = true
line-length = 120
indent-width = 4

target-version = "py310"

[tool.ruff.format]
preview = false
line-ending = "lf"
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 110
skip-magic-trailing-comma = false

[tool.ruff.lint]
fixable = ["ALL"]
ignore = []
preview = false
extend-ignore = []
extend-fixable = []
allowed-confusables = []
explicit-preview-rules = false
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
force-single-line = true
